+++ NEWS.md
@@ -1,6 +1,7 @@
 2020-xx-yy: dataiter 0.10
 =========================
 
+* `ListOfDicts.select`: Return keys in requested order
 * `Vector.__repr__`: Add custom conversion to string for display
 * `Vector.__str__`: Add custom conversion to string for display
 * `Vector.to_string`: Add custom conversion to string for display